// interfaces/barang.ts
export default interface IBarang {
  sku: string;
  nama_barang: string;
  kategori: string;
  group: string;
  brand: string;
  ukuran: string;
  kualitas: string;
  motif: string;
  tekstur: string;
  berat: number;
  gambar_url: string;
  design_url: string;
  gambar_design_url: string;
  external_sku: string;
  created_at: string;
  updated_at: string;
}